push

+++ src/main/java/com/driver/services/impl/AdminServiceImpl.java
@@ -46,9 +46,9 @@ public class AdminServiceImpl implements AdminService {
 	@Override
 	public void deleteAdmin(int adminId){
 		// Delete admin without using deleteById function
-
-		adminRepository1.deleteById(adminId);
-
+      if(adminId != 0) {
+		  adminRepository1.deleteById(adminId);
+	  }
 	}
 
 	@Override
